refactored dialogs into separate files, changed intents to use trigger actions, refactored azure code dialog flow
