+++ app/dialogs/botHelp.js
@@ -1,0 +1,14 @@
+const builder = require('botbuilder')
+
+const lib = new builder.Library('botHelp')
+lib.dialog('/', [
+  function (session, args, next) {
+    session.send('I can help you get an Azure Code, find resources, or connect you with our Microsoft team! What would you like help with?')
+  }
+]).triggerAction({
+  matches: 'botHelp'
+})
+
+module.exports.createLibrary = function () {
+  return lib.clone()
+}
