+++ app/dialogs/negativeComment.js
@@ -1,0 +1,14 @@
+const builder = require('botbuilder')
+
+const lib = new builder.Library('negativeComment')
+lib.dialog('/', [
+  function (session, args, next) {
+    session.send("That's not very nice.. If you come by the booth we can help you out in person!")
+  }
+]).triggerAction({
+  matches: 'negativeComment'
+})
+
+module.exports.createLibrary = function () {
+  return lib.clone()
+}
